+++ sync-for-payables/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v2.0.0] sync-for-payables
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/2.0.0 - sync-for-payables
+
+## 2023-12-14 14:07:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] sync-for-payables
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/2.1.0 - sync-for-payables